rewrote the express server to include routers
